Moved algo impl into separate class.
