+++ src/test/scala/org/economicsl/matching/MarriageMarketSpecification.scala
@@ -45,19 +45,15 @@ object MarriageMarketSpecification extends Properties("marriage-market") {
       .suchThat { case (ms, ws) => ms.size == ws.size }
   }
 
-  def isBlockedBy(matching: Matching[Woman, Man], pair: (Man, Woman)): Boolean = {
-    matching.get(pair._2).exists(pair._2.ordering.gt(pair._1, _)) && matching.get(pair._1).exists(pair._1.ordering.gt(pair._2, _))
-  }
-
   property("all men and women are matched") = forAll(unMatched) {
     case (ms, ws) =>
-      val (unMatchedMs, unMatchedWs, matching) = DeferredAcceptance.weaklyStableMatching(ms, ws)
+      val ((unMatchedMs, unMatchedWs), matching) = (new GaleShapleyAlgorithm[Man, Woman])(ms, ws)
       unMatchedMs.isEmpty && unMatchedWs.isEmpty && (matching.size == ms.size)
   }
 
   property("matching should be weakly stable") = forAll(unMatched) {
     case (ms, ws) =>
-      val (_, _, matching) = DeferredAcceptance.weaklyStableMatching(ms, ws)
+      val ((_, _), matching) = (new GaleShapleyAlgorithm[Man, Woman])(ms, ws)
       ms.forall(m => ws.forall(w => !matching.isBlockedBy(w -> m)))
   }
